Problem: ViewDB is still a vaporware
I've got stuck trying to implement indexing mechanics more
or less properly. However, some uncertainty, decision making
anxiety and procrastination ensured that I didn't move forward.

I've defined the following trait:

```
trait AttributeIndexingEnabled {
   "https://viewdb.org/attributes#factType" = "AttributeIndexingEnabled",
   "https://viewdb.org/attributes#attribute",
   "https://viewdb.org/attributes#closure"
}
```

I am not very happy about its shape. And the closure part was leaking
from the previous attempt to do this via generic triggers.

Also, it is not clear how it should work when we enable indexing after
covered attributes were already used; and what to do when we disable
and subsequently re-enable it. I mean, it is clear that everything that
is intended to be indexed, has to be indexed, but I feel like the semantics
of `AttributeIndexingEnabled` are not well thought through.

Something has to be done about this. I need to get unstuck.

Solution: drop indexing from the (unpublished) codebase altogether.

The idea here is to simplify the development process as much as
possible and get unstuck.

Sure, querying without indexing is going to be slow for a little
while. But indices are not THAT hard to bring back once things
start going.

My current hypothesis is that it is far more important to focus
on defining querying semantics and domain inference building blocks,
as those are the things that will make or break this approach.