Problem: query model is poorly defined
Well, technically speaking, it's not defined at all, but
rather a vague understanding of what it should be.

Solution: lay out the first draft of the query model

This adds viewdb_core and viewdb_query crates. The reason
for the split is to enable separate use of these in systems
that might (potentially) not be backed by PumpkinDB (which
is integrated into viewdb_engine)